chore: Platform test re-organization
